Split parts of `ide_db::call_info` off into `ide`
